conftest refactorization to use it only where is needed
