+++ .readthedocs.yaml
@@ -5,12 +5,12 @@ formats:
   - pdf
 
 sphinx:
-   configuration: docs/conf.py
+  configuration: docs/conf.py
 
 python:
   version: "3.7"
   install:
     - requirements: requirements.txt
-    - requirements: requirements-dev.txt  
+    - requirements: requirements-dev.txt
     - method: pip
       path: .
