--- enviornment.yml
@@ -1,10 +1,0 @@
-name: wefe
-dependencies:
-  - numpy
-  - scipy
-  - pandas
-  - scikit-learn
-  - gensim
-  - plotly
-  - six
-  - semantic_version
